backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
sqlalchemy==1.4.41
databases[postgresql]==0.6.2
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
python-multipart==0.0.20
requests==2.31.0
polyline==2.0.2
geopy==2.4.1